Created Individual_Reflection.tex
Created a .tex file for us to add our individual reflections in the report. The file has a template structure, the sections would just need to be filled in by yourselves. This is just an idea of how we can add our individual reflections, you may want to structure this section differently for yourself, which is totally fine :)